+++ spec/unit/installer/target_installer/aggregate_target_installer_spec.rb
@@ -157,8 +157,8 @@ module Pod
         )
         resources_by_config = @installer.send(:resources_by_config)
         resources_by_config.each_value do |resources|
-          resources.should.include '$CONFIGURATION_BUILD_DIR/BananaLib/Trees.bundle'
-          resources.should.include '$CONFIGURATION_BUILD_DIR/BananaLib/Leafs.bundle'
+          resources.should.include '$PODS_CONFIGURATION_BUILD_DIR/BananaLib/Trees.bundle'
+          resources.should.include '$PODS_CONFIGURATION_BUILD_DIR/BananaLib/Leafs.bundle'
         end
       end